Rename admin_grants.py to grants.py
